backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.2

# Added for security / config
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
